+++ packages/next/src/gqty/hooks/useLogout.ts
@@ -1,9 +1,10 @@
 import { config as coreConfig } from '@faustjs/core';
+// eslint-disable-next-line import/extensions
 import { LOGOUT_ENDPOINT_PARTIAL_PATH } from '@faustjs/core/config';
 import type { RequiredSchema } from '@faustjs/react';
 import { isNil } from 'lodash';
 import { useState } from 'react';
-import type { NextClient } from '../client';
+import type { NextClient } from '../client.js';
 
 export function create<
   Schema extends RequiredSchema,