+++ packages/next/src/gqty/hooks/usePreview.ts
@@ -1,8 +1,9 @@
+// eslint-disable-next-line import/extensions
 import { PageIdType, PostIdType } from '@faustjs/core/client';
 import type { RequiredSchema } from '@faustjs/react';
-import isUndefined from 'lodash/isUndefined';
-import { useRouter } from 'next/router';
-import { hasPageId, hasPostId } from '../../utils';
+import isUndefined from 'lodash/isUndefined.js';
+import { useRouter } from 'next/router.js';
+import { hasPageId, hasPostId } from '../../utils/index.js';
 import type { NextClientHooks, NextClientHooksWithAuth } from '.';
 
 export type UsePreviewResponse<Schema extends RequiredSchema> =