+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @faustjs/core
 
+## 0.15.1
+
+### Patch Changes
+
+- 3c1280b: Updated `schema.generated.ts` to support `contentNode` for `usePreviewNode` usage.
+
+## 0.15.0
+
+### Minor Changes
+
+- a044a07: Add file extensions to import statements to fully support ES Modules. Support for Next.js 12 🎉
+
 ## 0.14.1
 
 ### Patch Changes